Revert "Added Link State simulation"
