--- template/exercise.css
@@ -1,148 +1,0 @@
-/* General Body Styling */
-
-@import url('https://fonts.googleapis.com/css2?family=Poppins:wght@400;600&display=swap');
-
-@page {
-    size: A4;  /* You can also use other page sizes like 'letter', or custom dimensions */
-    margin: 10px;  /* Remove default margins */
-}
-
-
-body {
-    margin: 0;   /* Ensure no margins are added to the body */
-    padding: 0;  /* Remove padding if not needed */
-    width: 100%; /* Take full width of the page */
-    height: 100%; /* Take full height of the page */
-    border-radius: 8px;
-    background-color: #f1f1f1;
-    border: 1px solid #ddd;
-    font-family: 'Poppins', sans-serif;
-}
-
-
-/* Heading Styling */
-h1, h2 {
-    color: #333;
-    text-align: center;
-}
-
-.image-container {
-    text-align: center;
-    margin: 20px 0;
-}
-
-.image-container img {
-    max-width: 80%;
-    height: auto;
-    border-radius: 8px;
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-    align-items: center;
-}
-
-/* Exercise Instructions Styling */
-.exercise-instructions {
-    margin: 20px 5px;
-    background: #fff;
-    padding: 20px;
-    border: 1px solid #ddd;
-    border-radius: 8px;
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-
-}
-
-.exercise-instructions p, .exercise-instructions ul {
-    color: #555;
-    margin: 0;
-    padding: 0;
-}
-
-.exercise-instructions ul {
-    margin-top: 10px;
-    padding-left: 20px;
-}
-
-.exercise-instructions ul li {
-    margin: 8px 0;
-}
-
-/* Main Routing Messages Table Styling */
-.routing-messages {
-    margin-top: 20px;
-    margin: 10px;
-    border-radius: 8px;
-}
-
-.routing-tables {
-    margin-top: 20px;
-    margin: 10px;
-    border-radius: 8px;
-}
-
-table {
-    width: 100%;
-    border-collapse: collapse;
-    background: #fff;
-    border: 1px solid #ddd;
-
-    
-}
-
-th, td {
-    padding: 12px;
-    border: 1px solid #ddd; /* Adds separation lines between cells */
-    text-align: center;
-}
-
-th {
-    background-color: #f1f1f1;
-    color: #333;
-    position: sticky; /* Keeps header anchored */
-    top: 0;
-    z-index: 1;
-}
-
-td {
-    color: #555;
-}
-
-table caption {
-    font-size: 1.212em;
-    font-weight: bold;
-    text-align: center;
-    margin-bottom: 10px; /* Space between the title and the table */
-    margin-top: 20px;
-}
-
-
-.routing-messages th, .routing-messages td {
-    padding: 10px;
-    font-size: 14px;
-}
-
-/* Styling for Internal (Nested) Tables */
-.routing-messages table table {
-    width: 100%;
-    margin: 0;
-    border: 1px solid #ddd; /* Adds outer border to nested table */
-    border-collapse: collapse;
-    border-radius: 10px;
-}
-
-.routing-messages table table th,
-.routing-messages table table td {
-    border: 1px solid #ddd; /* Adds separation lines within nested tables */
-    padding: 8px;
-    font-size: 12px;
-    color: #666;
-    background: #fafafa;
-}
-
-/* Image Styling in Table */
-.routing-messages img {
-    width: 60px;
-    height: auto;
-    border-radius: 4px;
-    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-}
-
-
