--- template/exercise.html
@@ -1,103 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Routing Exercise</title>
-</head>
-<body>
-    <h1>Routing Exercise</h1>
-    <div class="image-container">
-        <img src="{{ image_path }}" alt="{{ image_path }}">
-    </div>
-
-    <div class="exercise-instructions">
-        <p>Consider the network in the figure and assume that:</p>
-        <ul>
-            <li>The adopted routing is {{ routing_protocol }}</li>
-            <li>The order of the generation of messages follows the numeric order of routers</li>
-            <li>The order of the arrival of messages follows the numeric order of routers</li>
-        </ul>
-
-        <p>Write down the list of generated messages and the final routing table. You can omit messages that are received but do not alter the routing table of the receiving router.</p>
-    </div>
-
-    <h2>Generated Messages</h2>
-    <div class="routing-messages">
-        <table>
-            <tr>
-                <th>Source Node</th>
-                <th>Destination Node</th>
-                <th>LSP Received</th>
-                <th>Routing Table</th>
-                <th>Network View</th>
-            </tr>
-            {% for message in generated_messages %}
-            <tr>
-                <td>{{ message['source'] }}</td>
-                <td>{{ message['destination'] }}</td>
-                <td>
-                    <table>
-                        <tr>
-                            <th>Neighbor</th>
-                            <th>Link Cost</th>
-                        </tr>
-                        {% for entry in message['lsp']['links'] %}
-                        <tr>
-                            <td>{{ entry['id_neigh'] }}</td>
-                            <td>{{ entry['cost'] }}</td>
-                        </tr>
-                        {% endfor %}
-                    </table>
-                </td>
-                <td>
-                    <table>
-                        <tr>
-                            <th>Path</th>
-                            <th>Cost</th>
-                        </tr>
-                        {% for entry in message['routing_table'] %}
-                        <tr>
-                            <td>{{ message['routing_table'][entry]['path'] }}</td>
-                            <td>{{ message['routing_table'][entry]['cost'] }}</td>
-                        </tr>
-                        {% endfor %}
-                    </table>
-                </td>
-                <td><img src="{{ message['routing_view'] }}" alt="{{ message['routing_view'] }}"></td>
-            </tr>
-            {% endfor %}
-        </table>
-    </div>
-        <h2>Final Routing Tables</h2>
-        {% for node in routing_tables %}
-            <div class="routing-tables">
-                <table>
-                    <thead>
-                        <tr>
-                            <th colspan="3" style="text-align: center; font-size: 1.2em;">Node {{ node }}</th>
-                        </tr>
-                        <tr>
-                            <th>Destination</th>
-                            <th>Path</th>
-                            <th>Cost</th>
-                        </tr>
-                    </thead>
-                    <tbody>
-                        {% for entry in routing_tables[node] %}
-                            <tr>
-                                <td>{{ entry }}</td>
-                                <td>{{ routing_tables[node][entry]['path'] }}</td>
-                                <td>{{ routing_tables[node][entry]['cost'] }}</td>
-                            </tr>
-                        {% endfor %}
-                    </tbody>
-                </table>
-            </div>
-        {% endfor %}
-
-
-                
-    </div>
-</body>
-</html>
